requirements.txt
typer==0.12.3
tabulate==0.9.0
python-dotenv==1.0.1
google-generativeai==0.7.2
